+++ src/adaptors/sommelier/v0-8-16.js
@@ -1,23 +1,11 @@
 const { default: BigNumber } = require('bignumber.js');
-const { request, gql } = require('graphql-request');
 const sdk = require('@defillama/sdk');
 const utils = require('../utils');
-const queries = require('./queries');
 const cellarAbi = require('./cellar-v0-8-16.json');
 const { chain } = require('./config');
+const { getApy } = require('./apy');
 
 const call = sdk.api.abi.call;
-
-// Use the change in price over 30 days to extrapolate an APR
-async function getApy(cellarAddress) {
-  const dayData = await queries.getDayData(cellarAddress, 30);
-
-  const price = new BigNumber(dayData[0].shareValue);
-  const prevPrice = new BigNumber(dayData[dayData.length - 1].shareValue);
-  const yieldRatio = price.minus(prevPrice).div(prevPrice);
-
-  return yieldRatio.times(12).times(100).toNumber();
-}
 
 // Call getPositions() to get a list of assets held by the Cellar
 async function getUnderlyingTokens(cellarAddress) {
